koly-health-diet-partner/scripts/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
torch==2.0.1
transformers==4.33.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
lxml==4.9.3
beautifulsoup4==4.12.2
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
nltk==3.8.1
spacy==3.6.1
datasets==2.14.4
accelerate==0.21.0
evaluate==0.4.0
